[전남대 3팀_BE] 3차 코드리뷰 (#51)
* feat: 관리자 페이지 전용 로그인 폼 구현 (#42)

* feat: 관리자 페이지 로그인 컨트롤러 및 Httponly 쿠키 방식으로 JWT 저장

* feat: 관리자 페이지 로그인 페이지 구현

* feat: JwtTokenFilter 책임 분리를 위한 JwtTokenResolver 구현

* feat: 정적 리소스 요청은 필터를 거치지 않게 수정

* feat: 관리자 페이지 관련 컨트롤러에 PreAuthorize 추가

* feat: 로그인 어노테이션 추가 (#43)

* feat: 로그인 어노테이션 및 ArgumentResolver 구현

* refactor: feed 컨트롤러에서 Login 어노테이션 사용하도록 수정

* refactor / service layer 부담 완화 (#45)

* feat : dto변환을 컨트롤러에게 넘기기

* feat : create와 update 모두 entity에서 책임지도록 변경

* feat : LazyInitializationException 방지

* feat: 피드 좋아요 기능 추가 (#46)

* feat: 피드 좋아요 엔티티인 FeedLike 추가

* feat: 피드에 좋아요 버튼을 누를 경우 반환할 FeedLikeResponseDto 추가

* feat: Feed 엔티티의 likeCount를 int에서 long으로 변경 및 likeCount 증감 메서드 추가

* feat: 피드 좋아요 토글 기능 추가

* refactor: feed 테이블의 like_count 타입을 BIGINT로 변경

* feat: 피드 좋아요 동시성 문제 해결을 위해 좋아요 수 증감은 DB레벨의 원자적 업데이트 사용하도록 변경

* fix: 피드 좋아요 관련 영속성 컨텍스트와 DB 불일치 해결

* feat: 배포, 개발 DB에 피드좋아요 테이블 추가

* feat: 피드에 좋아요를 누른 사용자 목록 조회 기능 추가

* feat: Swagger 적용 및 로그인 어노테이션 적용

* Update issue templates (#47)

* Update issue templates

* Update .github/ISSUE_TEMPLATE/전남대-3팀의-이슈-템플릿.md


---------


* feat: PR 템플릿 추가 (#48)

* feat: 프론트와 협업을 위한 임시 CORS 설정 (#50)

* refactor: 회원가입 시 프로필 이미지 필요 없게 수정

* feat: 임시 CORS 설정

* refactor: 배포 서버도 CORS 추가

---------

+++ src/main/java/com/starterpack/feed/entity/Feed.java
@@ -56,7 +56,7 @@ public class Feed {
     private Category category;
 
     @Column(name = "like_count", nullable = false)
-    private int likeCount = 0;
+    private long likeCount = 0;
 
     @OneToMany(mappedBy = "feed", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<FeedProduct> feedProducts = new ArrayList<>();
